fix a few things
